More struct to core, update for #32, imimage use struct
